startup-script.sh: Workaround for USB anomalies. 150MHz Dyplo clock
Set the Dyplo clock to 150MHz using devmem (ugly hack...)

With the USB hub in between, the touchscreen appears and then during
calibration disappears again because of some USB errors. This causes
the GUI to start without any pointer device at all.

To work around this, wait 4 seconds before starting. Spend the time
using readahead to fill the cache with the code we'll need later.