mandelbrotpipeline: Fix distortion when upper half is black
With preset C, the scheduler gets confused as the top lines take longer to calculate than
the rest of the frame, which causes some lines to be scheduled twice and results in
a distorted image.

Fix this by allowing up to 4 frames to be rendered ahead, so that workers can continue
working on more future frames. This solves the distortion, since the scheduler has much
more headroom now to keep feeding lines to nodes.