Allow MPLANE multiplanar capture capability
Allows to use devices that report multiple planes but actually
just use one plane. In particular, the Xilinx frame capture
device.

Add a workaround for the Xilinx device, it cannot really change
its resolution but tries and fails in an unexpected way. If a
driver fails the enumeration of resolutions, assume it's such
a buggy device and just use it at whatever resolution it was
already configured for.