Implement video cropping in the camera pipeline
If the camera renders much larger images than what we need, apply cropping
before processing the pixels. Start with vertical cropping only since
that is very easy to do, just move a pointer and change the size.
If the picture is much wider than the viewport, also implement horizontal
cropping.
This reduces the CPU load considerably when rendering a full-HD camera,
yielding 60 fps when rendered into a 640x480 viewport.

Also added a detach() call to the pixmap for rendering, make sure it's
not sharing data.