video-capture.cpp: Avoid the IMX274 1280x540 mode
The IMA274 has a lowest resolution of 1280x540, but this mode doesn't
have square pixels like most modes. Without any adjustments, the image
will appear "stretched" to the user, so just avoid this mode and pick
the 1280x720 one instead. If some other hardware happens to pick the
same mode, it'll also be avoided if possible.